+++ cynic-codegen/src/registration.rs
@@ -1,4 +1,5 @@
 use std::{
+    borrow::Cow,
     io::Write,
     path::{Path, PathBuf},
 };
@@ -46,14 +47,29 @@ impl<'a> SchemaRegistrationBuilder<'a> {
             name: &'a str,
             path: &Path,
         ) -> Result<SchemaRegistration<'a>, SchemaRegistrationError> {
-            let data = std::fs::read(path)?;
-            let registration = SchemaRegistration { name, data };
+            let data = std::fs::read_to_string(path)?;
+            let registration = SchemaRegistration {
+                name,
+                data: Cow::Owned(data),
+            };
             registration.write(registration.filename()?)?;
             registration.write_schema_module()?;
             cargo_rerun_if_changed(path.as_os_str().to_str().expect("utf8 paths"));
             Ok(registration)
         }
         inner(name, path.as_ref())
+    }
+
+    /// Registers a schema from a string of SDL
+    pub fn from_sdl(self, sdl: &'a str) -> Result<SchemaRegistration<'a>, SchemaRegistrationError> {
+        let SchemaRegistrationBuilder { name } = self;
+        let registration = SchemaRegistration {
+            name,
+            data: Cow::Borrowed(sdl),
+        };
+        registration.write(registration.filename()?)?;
+        registration.write_schema_module()?;
+        Ok(registration)
     }
 }
 
@@ -62,7 +78,7 @@ impl<'a> SchemaRegistrationBuilder<'a> {
 /// Additional methods can be called on this to
 pub struct SchemaRegistration<'a> {
     name: &'a str,
-    data: Vec<u8>,
+    data: Cow<'a, str>,
 }
 
 // Public API
@@ -87,29 +103,31 @@ impl SchemaRegistration<'_> {
         #[cfg(feature = "rkyv")]
         {
             filename.set_extension("rkyv");
-            let document_string = String::from_utf8(self.data.clone()).expect("schema to be utf8");
-            let document = crate::schema::load_schema(&document_string)
+
+            let document = crate::schema::load_schema(self.data.as_ref())
                 .map_err(|error| SchemaRegistrationError::SchemaErrors(error.to_string()))?
                 .into_static();
+
             let schema = Schema::new(SchemaInput::Document(document))
                 .validate()
                 .map_err(|errors| SchemaRegistrationError::SchemaErrors(errors.to_string()))?;
+
             let optimised = schema.optimise();
             let bytes = rkyv::to_bytes::<_, 4096>(&optimised).unwrap();
+
             Ok(std::fs::write(filename, &bytes)?)
         }
         #[cfg(not(feature = "rkyv"))]
         {
             filename.set_extension("graphql");
-            Ok(std::fs::write(filename, &self.data)?)
+            Ok(std::fs::write(filename, self.data.as_bytes())?)
         }
     }
 
     fn write_schema_module(&self) -> Result<(), SchemaRegistrationError> {
         use crate::use_schema::use_schema_impl;
 
-        let document_string = String::from_utf8(self.data.clone()).expect("schema to be utf8");
-        let document = crate::schema::load_schema(&document_string)
+        let document = crate::schema::load_schema(self.data.as_ref())
             .map_err(|error| SchemaRegistrationError::SchemaErrors(error.to_string()))?
             .into_static();