+++ CHANGELOG.md
@@ -9,6 +9,10 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ## Unreleased - xxxx-xx-xx
 
+### New Features
+
+- Added `from_sdl` method to schema registration.
+
 ## v3.0.0-beta.2 - 2023-05-13
 
 ### Bug Fixes
@@ -25,7 +29,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - The workings of `derive(QueryVariables)` has been changed quite
   significantly, but shouldn't affect anyone who is using the macro.
 - The various cynic derives now expect the crate to be in scope under the name
-  `cynic`.  This hopefully shouldn't affect most users but may do so if you
+  `cynic`. This hopefully shouldn't affect most users but may do so if you
   were doing something unusual.
 
 ### New Features